requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-Sock==0.7.0
boto3==1.28.44
requests==2.31.0
pypdf==3.16.0
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
gunicorn==21.2.0
celery[redis]==5.3.4
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0
docx2txt==0.8
python-docx==0.8.11
textract==1.6.5
pydantic==2.4.2
redis==4.5.4
flower==2.0.1
langchain==0.0.300
langchain-openai==0.0.2
fastapi==0.103.1
uvicorn==0.23.2
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
isort==5.12.0
flake8==6.1.0
httpx==0.27.2 